run Ink compiler as a part of Vite build
