+++ package.json
@@ -4,13 +4,11 @@
   "license": "MIT",
   "type": "module",
   "scripts": {
-    "prestart": "node ink-compile.cjs",
     "start": "vite",
-    "prebuild": "node ink-compile.cjs",
     "build": "vite build",
     "preview": "vite preview",
     "lint": "eslint src",
-    "singlefile": "node ink-compile.cjs && vite build --mode singlefile"
+    "singlefile": "vite build --mode singlefile"
   },
   "dependencies": {
     "@atrament/web": "1.2.3",
